Revert changes to WrappedCursor.to_list
Motor never supported passing callbacks to to_list when using asyncio,
so we aren't actually gaining any compatibility by trying to enable it
here. Instead, adjust how we call to_list() to only use the callback
keyword argument if a callback was passed in. This will cause motor to
throw the appropriate Exception based on which version is installed.